Break out remaining helper classes into their own files
